Renamed variables in forth.c, change div / 0
The variables in "forth.c" have been renamed, which looks like a big change,
but it was just a search and replace.

The behavior of the virtual machine on a division by 0 has been changed, it
sets all relevant registers to '10' and jumps to the location 1, which the
image has set to '-throw'.