Improved (save), more commands line options, ...
The "(save)" instruction now accepts a range, although it still needs testing
properly, this will be used by the metacompiler to generate new images.

Some To Do comments have been updated in the read me file.

The metacompiler assembler is almost working, however there are problems with
how the vocabulary works, mainly due to how ";" and ":" are implemented - which
could be fixed with a proper word smudging mechanism, or by other methods. It
is making doing certain things more convoluted than necessary.

Command line parsing has been improved so new images can be named and saved if
needed, this is reflected in the makefile.