Added more code to the metacompiler
'dump' now works in the metacompiler. Some more words are need to accommodate
the current code, such as target only words (or the vocabulary will have to be
tightly controlled). For example 'literal' is used by the metacompiler, and is
a word in the target, as well as being a normal Forth word - the correct one
will have to be used.