Metacompiler fully working!
The metacompiler and metacompiled program are fully working, and a target in
the makefile has been added to fix this. This required 'immediate' and 'inline'
to be implemented in the metacompiled program and the built in primitives to have
their inline bits set as well as a 'nop' inserted.

What is left is: optimizations in the metacompiler, documentation, cleanup and
removing the C version of the compiler but leaving a metacompiled image in the
repository.