Smaller image size, more error checking, file format
The image size is now smaller, but there is still a lot more to do in order to
shrink the program.

More error checking has been added to the metacompiler so mistakes can be
caught.

At the beginning of the generated image a section detailing the file format
has been added, this includes file format, endianess, CRC and length fields.

The 'bi' script has been tailored for 'meta.fth'.