Many more documentation changes, reorganization
Some of the code has been reorganized so words are grouped together in a more
logical order, and 'abort"' has been fixed so it is now standards compliant (it
checks for a non-zero value before throwing an abort and printing a string).

As is usual now, much more documentation has been added, some sections of code
have nearly been completely documented.