5502: Added temporary code demonstrations
Some temporary code has been added, which will be removed in the next commit,
which demonstrates a few different things possible with this project. A new
target has been added called 'eforth' which contains an complete working eforth
image ready to go and built into the generated executed. It does this by
turning 'eforth.blk' into a C file (called 'eforth.gen.c', generated with
'blk2c.c') and building 'eforth.c' and 'embed.c' with the compile time switch
'-DINCLUDE_DEFAULT_IMAGE'. Another demonstration includes the 'custom-embed'
directory, which includes a hacked 'embed' interpreter which allows evaluation
and saving a running machine and restoring it to its *exact* state it was when
it was saved.

All traces of these things will be removed.