5510: Moved registers to core
The registers in the virtual machine are now saved in the header for virtual
machine image, so the initial stack pointers, program counter and top of stack
can now be specified instead of being set. This required some reengineering of
the virtual machine, image and meta compiler. It has also revealed a bug which
causes the metacompilation toolchain to fail if the header varies in size (the
exact circumstances and reasons for failure elude me at the moment, if the
header stays the same size, everything is fine).