5414: added um/mod as virtual machine instruction
'um/mod' has been added as a virtual machine instruction, replacing 'u/mod'.
This means that 'um/mod' does not have to be implemented in Forth, which is
best for efficiency reasons. The code for a software 'um/mod' implemented in
terms of 'um+' is left in there. '/mod' is left unchanged, which is not
symmetrical as it only operates on signed single cells values, unlike 'um/mod'
which operates on a double and a single cell. This is in practice more useful.

There are other minor changes in the way things are loaded off the disk, but
they are of little consequence.