5378: Simplified digit handling, and 'see'
Apart from some other minor changes and comments, the two big changes in this
commit is how digit input handing is done and a minor improvement to 'see'.
Numeric input has been simplified but the functions and how they work stay the
same.

There has been a minor improvement to 'see', which could actually use a lot of
work. 'see' now attempts to decompile branches into jumps to known words, which
occurs when a call to a word followed by an exit gets turned into a tail call.

'd+' has been added as a word as well.