5424: 'yield?', tracing, simplification
The way parsing works has been simplified, it is now not reliant on a 'test'
variable execution vector. As always things have been made to be smaller and
denser where possible.

'yield?' has been added, which can give more control over the save/restore
state of the virtual machine. It perhaps needs more thinking about.

A register that enables tracing of the virtual machine as it executes has been
added, a word call 'trace' has been added which enables this tracing feature
for the duration of a words execution.