5350: Fix DPL, VM, tests, ...
A bug with the double cell numeric input handling has been fixed, the number
was placed on the stack back to front.

The Virtual Machine has had minor changes, an 'assert' was replaced with an
exit code and some of the return logic was changed.

Example code is being added to 'unit.fth', which is a good place for it. Code
for low level bit manipulation routines, stack and string manipulation has been
added from various named sources.

Some of the Forth block editor commands have been changed so the names do not
clash with hexadecimal values.