5350: Adding in test code, fixed 'um/mod' bug!
Some test code has been added for floating point arithmetic, it was previously
in the project but has been added again, this time it appears to be working
correctly after fixing a bug with 'um/mod'. Some of the previous programs have
been added back in, they could be made to be part of the appendix in the book
generated from 'meta.fth'.