5310: Added optional image obfuscation, error code
It is now possible to create an 'encrypted' image, one that is obfuscated
making analysis of the generated binary more difficult. It is not a difficult
scheme to crack, and is more a demonstration that work can be done in the
boot-loader that manipulates the rest of the image in non-trivial ways.

The 'embed.c' virtual machine now returns Forth error codes if things go wrong
instead of dieing in more circumstances as well.