5290: Trace, callbacks added to API, default, status
There are a fair few changes in this commit to how things work:

* The callbacks for the default functionality to save an image, get a character
and write a character are now accessible via the API.
* A default state has been added to the VM whereby an exception is raised if
an unknown ALU instruction is executed.
* 'dist' is now a makefile target, which will make a compressed TAR archive
of the relevant project files for a release ('install' and 'uninstall' are
not targets as of yet).
* The manual page still needs work, but has been improved.
* The CPU status register can now be set by an option in the option struct,
this could be used to communicate to the eForth image if we are reading from
standard in, or if we are using blocking calls, or whether to display the
welcome message an okay prompt.
* Some @todos have been added