5332: Terminal handling options added
Terminal handling options have been added, and can be selected for via the
options flags passed into the virtual machine. This allows the device to talk
over a serial port if needed. This is simulated with a test program 'Unix',
which turns off blocking while reading from the standard input file descriptor,
and puts the terminal into raw mode (no line buffer done by the terminal).

* The documentation will need adjusting for this change.
* Fix echoing behavior for Unix test program
* Get a Windows equivalent working
* Add a silent (that disables '<ok>' and the welcome message) and verbose mode
* Improve handling of start/stop conditions for the Virtual Machine,
PC, SP, RP and T registers get overwritten, instead a shadow copy
could be initialized from if an option is specified (or not). This would allow
us to do other things and stop and start the interpreter as we like.