5376: Improved API documentation greatly
The API documentation in 'embed.h' has been vastly improved, with nearly all
the function fully documented, examples are missing which could be added in.
The manual page has also been improved.

The '--' option has been added.

The fgetc callbacks now can return a blocking status if needed, although the
application does not as of yet take advantage of it.