5366: MMU functionality, moved variables, windows
* A windows test program has been added called 'win.c' that does the same as
'unix.c', but for Windows. The 'call.c' program has also had its typedef for
floats changed from 'float_t' to 'vm_float_t' as there was a conflict with one
of the headers.
* 'MMU', or 'Memory Management Unit' functionality has been added, which are a
pair of functions that are used to load or store memory to locations in the
Virtual Machine, this allows logic to be written that handles accessing those
locations - such as mapping sections of memory to Read Only Memory (ROM), or
debugging.
* 'embed_puts' has been added to the API
* Using the new MMU mechanism all writes to memory locations were logged, after
some slight changes and moving some variables around it turns out only about
200 bytes are needed to implement the command interpreter and most words
(except any defining words).