5378: win.c, disassembler, word headers
'win.c' has been updated to work with file redirection as well as the command
prompt.

A disassembler has been added to 'embed.c' which will disassemble instructions
in a limited fashion. It is quite compact and can be enabled as part of the
trace output, it generates a table of the decompiled instruction.

The major, partially completed change, is that the word header bits for
immediate and compile-only words has been moved from the previous word pointer
field and into the word length field, which makes traversal of the dictionary
easier (and consequently manipulation of the word name field more difficult).

Everything nearly works with the word header change, the Forth decompiler
'see' however has a bug where by it cannot find defined words making the
decompiler less useful.

The documentation for the word header changes needs to be added as well.

The changes are on going.