5368: Documentation, non-blocking, 'unix.c'
* 'unix.c' has been updated
* The shadow register option has been removed, instead 'embed_reset()' can be
used instead.
* The documentation has beep updated to reflect recent changes in terms of the
bits used in the word header for 'compile-only' and 'immediate' words.
* The biggest, compatibility breaking, change is a proper mechanism for
non-blocking operation is now built into the 'rx?' ALU instruction in the VM.
This might need tweaking, but it works. The user should not have to specify
whether non-blocking would take place as the 'fgetc' callback can supply
whether it there is any data or not. This needs to be removed as an option.