5362: Massive API reorganization
The API for the embed library has been massively reorganized, functions that
will be more difficult to port to an embedded platform like logging and
allocation have been mostly been moved into a 'util.c' file which is not built
into the library. Also, two new callbacks have been added for reading and
writing to the virtual machines memory, this can be used to map sections into
ROM or have memory mapped IO, it would also be possible to have empty sections
to save on memory.

Other changes include:

* The way the options structure gets passed into the run time has changed, a
copy of the structure is now stored in 'embed_t'.
* An example program has been added to demonstrate the new MMU callbacks
* The test programs and drivers have been updated to reflect the new API (the
Windows test program has not been tested, it is likely to not compile).