5380: Modified workings of the BYE instruction
The BYE instruction used in 'yield?' has been changed so now only one
parameter is left on the stack when the virtual machine yields, which
makes for a slightly less confusing API when 'embed_pop' and
'embed_push' are used after calling 'embed_eval' in a C program. This
situation is still not satisfactory and will be improved.

Also of note, the test driver program in 'main.c' that provides the
Command Line Program for the embed library now does not allocate an
'embed_t' structure, but instead uses static allocation. This is part of
an ongoing effort to make the library more suitable for porting to a
microcontroller.