5408: Improved 'see', still buggy, also 'yield!?'
'see' has been improved, it is still buggy when it comes to decompiling
words in different vocabularies (the stack parameters are not being
managed correctly), but at least it does not get stuck and carry on
printing forever.

'yield!?' has been added to managed some of the complexities of the raw
'yield?' instruction - it is an odd instruction, but it works.