5414: added 'f2md.fth', makefile targets, b2c.fth
* Some changes in 'b2c.fth' have been made, only relating to the
messages printed when the program has been compiled.
* A program called 'f2md.fth' has been created which can be used to
create a filter program, stored as a block, called 'f2md.fth'. This will
be used to replace the AWK script, in 't/convert'. The script converts a
specially formatted Forth code into a markdown file, which can be
converted into HTML/PDF or other presentation formats. It is partially
completed (escaping is missing, and some other things).
* makefile targets have been added for the conversion program.