5414: Corrected typo, updated 'readme.md'
A typo in the man page has been corrected (the man page should not
really exist and instead should be generated from the 'readme.md' file,
like some of my other projects, but that is another story).

The 'readme.md' file has been updated to refer to another branch I am
working on that simplifies the project and makes it more like the H2 CPU
written in VHDL.

As an aside: Perhaps the executable produced should be called 'eforth'
instead of 'embed'?